# Replace ByteArrayBuilder with Vec