Replaces `ByteArrayBuilder` with a plain old `Vec`.

The goal was faster text pool creation. I thought avoiding zeroing out the vec would be more significant. But there's almost no change. I do think using a `Vec` and reducing some of the abstraction around this can help future optimization work though.

Vast majority of the time is still spent on `memmove`. I think there's a path towards doing fewer larger batches for `memmove` instead of the many small calls we're doing now. But that'd need to be in a followup at some point.

Perf changes within noise threshold. Ran three times with each branch with small sf to test text pool initialization.

```
time target/release/tpchgen-cli -s 0.01 --output-dir=/tmp/tpchdbgen-rs
```

| main | this branch |
|-|-|
| 1.017 secs, 1.022 secs, 1.023 secs | 1.008 secs, 1.019 secs, 1.013 secs|